type cast in java

+++ java/DSA/src/area_of_circle_8.java
@@ -3,8 +3,8 @@ public class area_of_circle_8 {
     public static void main(String[] args) {
         System.out.println("enter the radius of circle : ");
         Scanner sc=new Scanner(System.in);
-        int radius = sc.nextInt();
-        double pai = 3.14;
+        float radius = sc.nextFloat();
+        float pai = 3.14f;
         System.out.println("area of circle is :"+(pai*radius*radius));
         sc.close();
     }
